refactor: eliminate object/Any — introduce PydanticJson boundary layer
Replace all dict[str, object], Any, and untyped generic dicts throughout
the codebase with a strict two-zone type system:

  Internal code  → TypedDicts + JSONValue (mypy-only, no Pydantic)
  Pydantic fields → PydanticJson (RootModel, schema-generation safe)

New module app/contracts/pydantic_types.py:
  PydanticJson    — named recursive RootModel; the only safe recursive
                    JSON type in Pydantic v2 BaseModel fields
  wrap/unwrap     — single boundary converters; no cast() at call sites
  wrap_dict/unwrap_dict — dict[str, JSONValue] ↔ dict[str, PydanticJson]

New Pydantic wire models for self-referential structures:
  ToolCallWire         — CompleteEvent.tool_calls + ToolCallEvent.params
  MCPPropertyDefWire   — mirrors MCPPropertyDef (self-ref via model_rebuild)
  MCPInputSchemaWire   — mirrors MCPInputSchema
  MCPToolDefWire       — mirrors MCPToolDef; used as FastAPI response type

json_list() gains @overload signatures for NoteDict/CCEventDict/PitchBendDict
to resolve list invariance without cast() at any call site.
jint()/jfloat() added for safe JSONValue numeric extraction.

GenerationStep.constraints, ProtocolEventsResponse.events,
ProtocolSchemaResponse.events, and ProtocolToolsResponse.tools updated
to PydanticJson/MCPToolDefWire to fix RecursionError in schema generation.
health.py TypedDict import fixed for Pydantic v2 on Python < 3.12.

Golden hash updated to 424dca95. All tests pass. Both mypy targets clean.

+++ tests/test_neural_mvp.py
@@ -11,7 +11,7 @@ Tests:
 from __future__ import annotations
 
 import pytest
-from app.contracts.json_types import NoteDict
+from app.contracts.json_types import NoteDict, jfloat, jint
 from app.core.emotion_vector import (
     EmotionVector,
     emotion_to_constraints,
@@ -428,7 +428,7 @@ class TestHuggingFaceBackend:
         low_params = backend._emotion_to_hf_params(low_energy, bars=4)
         
         # Higher energy/tension should mean higher temperature
-        assert high_params["temperature"] > low_params["temperature"]
+        assert jfloat(high_params["temperature"]) > jfloat(low_params["temperature"])
     
     def test_emotion_to_params_motion_affects_tokens(self) -> None:
 
@@ -441,7 +441,7 @@ class TestHuggingFaceBackend:
         high_params = backend._emotion_to_hf_params(high_motion, bars=4)
         low_params = backend._emotion_to_hf_params(low_motion, bars=4)
         
-        assert high_params["max_tokens"] > low_params["max_tokens"]
+        assert jint(high_params["max_tokens"]) > jint(low_params["max_tokens"])
     
     def test_emotion_to_params_intimacy_affects_top_p(self) -> None:
 
@@ -455,7 +455,7 @@ class TestHuggingFaceBackend:
         distant_params = backend._emotion_to_hf_params(distant, bars=4)
         
         # More intimate = more focused = lower top_p
-        assert intimate_params["top_p"] < distant_params["top_p"]
+        assert jfloat(intimate_params["top_p"]) < jfloat(distant_params["top_p"])
     
     @pytest.mark.asyncio
     async def test_fallback_when_no_api_key(self) -> None:
